requirements.txt
anyio==3.6.1
asgiref==3.5.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.2
black==22.3.0
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
distlib==0.3.3
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
filelock==3.1.0
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
virtualenv==20.8.1
